* Reorganization of Graphs folder * Break graph.scss into smaller files cor movable-line, movable-value, axis and marquee
