+++ v3/src/components/app.tsx
@@ -5,7 +5,7 @@ import {Container} from "./container"
 import {DataSummary} from "./data-summary"
 import {gDataBroker} from "../data-model/data-broker"
 import {DataSet, toCanonical} from "../data-model/data-set"
-import {Graph} from "./graph/graph"
+import {Graph} from "./graph/components/graph"
 import {Text} from "./text"
 import { dndDetectCollision } from "./dnd-detect-collision"
 import {useDropHandler} from "../hooks/use-drop-handler"
