+++ v3/src/components/graph/components/axis.scss
@@ -1,0 +1,41 @@
+.axis {
+  background-color: white;
+
+
+  .axis-title {
+  }
+
+  .dragRect {
+    fill: #e38e71;
+    fill-opacity: 0.01;
+    cursor: grab;
+  }
+
+  .dragging {
+    cursor: grabbing !important;
+  }
+
+  .h-translate {
+    cursor: grab;
+  }
+
+  .h-upper-dilate {
+    cursor: url("../../../assets/RightDilate.png") 8 8, grab;
+  }
+
+  .h-lower-dilate {
+    cursor: url("../../../assets/LeftDilate.png") 8 8, grab;
+  }
+
+  .v-translate {
+    cursor: url("../../../assets/TranslateY.png") 8 8, grab;
+  }
+
+  .v-upper-dilate {
+    cursor: url("../../../assets/UpDilate.png") 8 8, grab;
+  }
+
+  .v-lower-dilate {
+    cursor: url("../../../assets/DownDilate.png") 8 8, grab;
+  }
+}
