+++ v3/src/components/graph/components/graph.scss
@@ -1,0 +1,45 @@
+.graph-plot {
+  width: 100%;
+  height: 100%;
+  background-color: aliceblue;
+}
+
+.graph-svg {
+  width: 100%;
+  height: 100%;
+  margin: 25px;
+}
+
+.graph-plot text {
+  fill: black;
+  font: 12px sans-serif;
+}
+
+.graph-dot-area {
+  background: white;
+}
+
+/**
+todo: Generally, .scss should be as close to its code as possible, which generally means associated with particular components
+where possible, and often nested within each other which minimizes the chances of global conflicts. That way, if the
+case table also wants to use .background as a class name the two definitions won't conflict with each other.
+ */
+.graph-background {
+  fill: white;
+}
+
+.graph-dot {
+  fill: rgba(232, 147, 118, 0.85);
+  stroke: white;
+  cursor: pointer;
+}
+
+.graph-dot-highlighted {
+  fill: steelblue;
+}
+
+.graph-plot-choice {
+  bottom: 10px;
+  right: 10px;
+  position: sticky;
+}
