+++ v3/src/components/graph/components/marquee.scss
@@ -1,0 +1,14 @@
+// Marquee classes
+.marquee {
+  fill: none;
+  stroke-width: 2;
+  stroke: gray;
+  stroke-dasharray: 5;
+}
+
+.marquee-back {
+  fill: none;
+  stroke-width: 2;
+  stroke: white;
+}
+
